notes/06_python_agents/pyproject.toml
[project]
name = "06-python-agents"
version = "0.1.0"
description = "Google ADK to create agents"
authors = [
    {name = "BobasB",email = "bugil.bogdan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "google-adk (>=1.26.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
